Copied target source for testing into test properties folder.
This is to ensure that our tests do not suddenly fail because the
content of the page source has changed. Also, it was tempting to
refactor so that we directly used Jsoup's read from URL, but I felt it
more important to retain the ability to inject any String into the
PageProcessor method, for testability.