Refactored initial PageReader tests to verify content.
Since all our subsequent tests assume the page-source.html is correct,
it makes sense to check that is still the case. If this source suddenly
changes, the tests will fail, bringing the change to our attention. It
may just require an update of page-source.html, but it also may indicate
legit refactoring of PageProcessor is required.